# src/swagger.yaml
openapi: 3.0.3
info:
  title: Worldwide Radio API (API de Emissoras de Rádio Mundiais)
  version: 1.0.0
  description: |
    API REST global para consulta, pesquisa, filtros e streaming de emissoras de rádio de todo o mundo.
    Combina um catálogo curado de alto rendimento com acesso à rede global com mais de 40.000 estações.
servers:
  - url: /api
    description: Servidor da API
paths:
...
  /radios/{id}/stream:
    get:
      summary: Transmitir áudio da estação (302 Redirect)
      description: Redireciona o reprodutor diretamente para o URL do stream de áudio ativo.
      parameters:
        - name: id
          in: path
          required: true
          description: ID ou UUID da estação
          schema:
            type: string
      responses:
        '302':
          description: Redirecionamento direto para a transmissão de áudio.
        '404':
          description: Fluxo indisponível ou rádio não encontrada.
  /radios/{id}/proxy:
    get:
      summary: Proxy de áudio resiliente
      description: Retransmite o áudio diretamente pelo servidor com conexão keep-alive sem timeout e headers CORS.
      parameters:
        - name: id
          in: path
          required: true
          description: ID da rádio
          schema:
            type: string
      responses:
        '200':
          description: Fluxo de áudio contínuo.
...